Pre-Deploy

+++ core/views/generic.py
@@ -1,9 +1,7 @@
-import bcrypt
 from flask import request, render_template, redirect
 from flask.views import View
-from catalog.models import User
-from catalog.session import session
 from flask import session as a_session
+from catalog.session import session
 
 
 class TemplateView(View):
@@ -46,17 +44,13 @@ class AuthorizedView(SingleObjectView):
 
     def check_permissions(self):
         user = a_session.get('user')
-        if user:
-            return True if self.object.created_by == float(user['id']) else False
+        if user and self.object.is_owner(user):
+            return True
         else:
             return False
 
     def dispatch_request(self, **kwargs):
-
-        if self.check_permissions():
-            return self.render_template()
-        else:
-            return self.render_error(message='Unauthorized Request')
+        return self.check_permissions()
 
 
 class IndexView(TemplateView):
@@ -67,45 +61,6 @@ class IndexView(TemplateView):
             return redirect('/categories')
         else:
             return self.render_template()
-
-
-class LoginView(View):
-    methods = ['GET', 'POST']
-    redirect_url = '/categories/'
-    template_name = 'login.html'
-
-    @staticmethod
-    def get_user():
-        return session.query(User).first()
-
-    def get_template_name(self):
-        raise NotImplementedError()
-
-    def render_template(self, context={}):
-        return render_template(self.get_template_name(), **context)
-
-    def authenticate(self, username, password):
-        # TODO: Check creds with DB
-        hashed_password = b'$2b$12$Fll4TLljj7dYmJMeC5jb2e5ilmj4JGpkbkiMeQo7tk.vNCe6HVewi'
-        if bcrypt.checkpw(password.encode('utf-8'), hashed_password):
-            a_session['username'] = username
-            a_session['logged_in'] = True
-            return True
-
-    def dispatch_request(self):
-        if request.method == 'POST':
-            if self.authenticate(request.form.get('username'), request.form.get('password')):
-                return redirect(self.redirect_url)
-        else:
-            return self.render_template()
-
-
-class LogoutView(TemplateView):
-    template_name = 'logout.html'
-
-    def dispatch_request(self):
-        a_session.clear()
-        return self.render_template()
 
 
 class ListView(TemplateView):
@@ -147,15 +102,18 @@ class CreateView(TemplateView):
 
 class UpdateView(AuthorizedView, TemplateView):
 
-    def process_form(self, *args, **kwargs):
-        o = session.query(Item).first()
-        print(o.id)
-        o.name = 'JPL'
+    def process_form(self, form, *args, **kwargs):
+        self.object.name = form['name']
+        self.object.price = form['price']
+        self.object.condition = form['condition']
+        self.object.production_year = form['production_year']
+        self.object.category_id = form['category_id']
         session.commit()
 
     def dispatch_request(self, **kwargs):
         self.get_context()
-        self.context['object'] = self.get_object(pk=kwargs['pk'])
+        self.get_object(pk=kwargs['pk'])
+        self.context['object'] = self.object
         # Handle Form Update
         if request.method == 'POST':
             self.process_form(form=request.form)
@@ -163,14 +121,12 @@ class UpdateView(AuthorizedView, TemplateView):
         # Show Form to owner
         else:
             if request.method == 'GET':
-                super(UpdateView, self).dispatch_request(**kwargs)
-                # Show Form to object owner
-                return self.render_template()
-            else:
-                # Show Error to other users
-                return self.render_error('Unauthorized Request')
-
-
+                if super(UpdateView, self).dispatch_request(**kwargs):
+                    # Show Form to object owner
+                    return self.render_template()
+                else:
+                    # Show Error to other users
+                    return self.render_error('Unauthorized Request')
 
 
 class DeleteView(AuthorizedView):
@@ -184,9 +140,11 @@ class DeleteView(AuthorizedView):
 
         self.get_object(pk=kwargs['pk'])
 
-        super(DeleteView, self).dispatch_request(**kwargs)
-        try:
-            self.delete_object()
-            return redirect('/categories/{}'.format(kwargs['category_pk']))
-        except Exception as e:
-            return self.render_error('Error while trying to delete item')
+        if super(DeleteView, self).dispatch_request(**kwargs):
+            try:
+                self.delete_object()
+                return redirect('/categories/{}'.format(kwargs['category_pk']))
+            except Exception as e:
+                return self.render_error('Error while trying to delete item')
+        else:
+            return self.render_error('Unauthorized Request')
